make changes asked by the CR

+++ app/controllers/recipe_foods_controller.rb
@@ -1,31 +1,48 @@
 class RecipeFoodsController < ApplicationController
-  def create
-    recipe_food = RecipeFood.new(recipe_food_params)
-
-    if recipe_food.save
-      flash[:success] = 'Food added successfully'
-    else
-      flash[:error] = 'There was an issue adding the food'
+  def new
+    recipe = Recipe.find(params[:recipe_id])
+    unless recipe.user == current_user
+      flash[:alert] =
+        'You do not have access to add an ingredient on a recipe that belongs to other Users.'
+      return redirect_to recipes_path
     end
 
-    redirect_to recipe_path(recipe_food_params['recipe_id'])
+    @recipe_food = RecipeFood.new
+    @foods = current_user.foods
+  end
+
+  def create
+    @recipe = Recipe.find(params[:recipe_id])
+    recipe_food = RecipeFood.new(recipe_food_params)
+    recipe_food.recipe = @recipe
+
+    if recipe_food.save
+      redirect_to recipe_path(@recipe), notice: 'New ingredient was successfully added.'
+    else
+      flash[:alert] = 'New Ingredient adding Failed. Please try again.'
+    end
   end
 
   def destroy
     recipe_food = RecipeFood.find(params[:id])
 
-    if recipe_food.destroy
-      flash[:success] = 'Ingredient deleted successfully'
-    else
-      flash[:error] = 'There was an issue deleting the ingredient'
+    unless recipe_food.recipe.user == current_user
+      flash[:alert] =
+        'You do not have access to delete the ingredient belongs to other Users.'
+      return redirect_to recipes_path
     end
 
-    redirect_to recipe_path(recipe_food.recipe)
+    if recipe_food.destroy
+      flash[:notice] = 'Ingredient was successfully removed.'
+    else
+      flash[:alert] = 'Ingredient removal Failed. Please try again.'
+    end
+    redirect_back(fallback_location: root_path)
   end
 
   private
 
   def recipe_food_params
-    params.require(:new_recipe_food).permit(:quantity, :food_id, :recipe_id)
+    params.require(:recipe_food).permit(:quantity, :food_id)
   end
 end
